+++ include_private/jsdrv_prv/thread.h
@@ -30,6 +30,7 @@
 #else  /* presume POSIX */
 #include <pthread.h>
 #endif
+#include <stdbool.h>
 
 /**
  * @ingroup jsdrv_prv
@@ -63,6 +64,7 @@ typedef pthread_t jsdrv_thread_t;
 
 int32_t jsdrv_thread_create(jsdrv_thread_t * thread, jsdrv_thread_fn fn, THREAD_ARG_TYPE fn_arg);
 int32_t jsdrv_thread_join(jsdrv_thread_t * thread, uint32_t timeout_ms);
+bool jsdrv_thread_is_current(jsdrv_thread_t * thread);
 void jsdrv_thread_sleep_ms(uint32_t duration_ms);
 
 JSDRV_CPP_GUARD_END